perf: cut install size 42% and web entry chunk 32%
CLI / packages (install: 122 MB -> 71 MB of node_modules):
- remove effect from template-generator: preflight-validation.ts rewritten
  in plain TypeScript (drops effect 33 MB + fast-check 5 MB)
- remove unused memfs dep (drops @jsonjoy.com ~15 MB); virtual-fs was
  already a plain Map
- remove dead CLI deps: handlebars, tinyglobby, ts-morph, yaml (all
  unimported in apps/cli/src)
- stop scaffolding deprecated @effect/schema: Schema ships in effect core
  since 3.10 (effect-full and --validation effect-schema now add effect
  only); display labels updated to "Effect Schema"
- lazy-load the template-generator bundle (2.5 MB of embedded templates)
  off the CLI startup path: bin entry now imports run.ts and handlers
  dynamically import the generator at point of use; index.ts keeps the
  full programmatic API unchanged

Web (entry chunk 438 KB -> 298 KB gzip, docs content fully code-split):
- docs/guides MDX modules and raw markdown are now lazy per-page chunks;
  frontmatter comes from a build-time virtual module
  (vite-plugins/content-meta) so the entry no longer embeds the whole
  docs site twice (compiled + ?raw)
- docs search index builds on first open from lazy raw chunks
- navbar no longer drags stack-translation/compatibility/TECH_OPTIONS
  into the entry: render-time slug check moved to a dependency-free
  module, copy-command parsing loads at click time
- drop root ConvexProvider (no routed page consumes convex)
- Caveat font moved from a render-blocking CSS @import to head links
  with preconnect (also fixes the @import-order build warning)
- refresh perf-baseline.json (totalJsGzip rises from per-page chunk
  overhead, but per-page download drops sharply)
