+++ apps/cli/package.json
@@ -137,15 +137,11 @@
     "execa": "^9.6.1",
     "fs-extra": "^11.3.5",
     "gradient-string": "^3.0.0",
-    "handlebars": "^4.7.9",
     "jsonc-parser": "^3.3.1",
     "oxfmt": "^0.19.0",
     "picocolors": "^1.1.1",
-    "tinyglobby": "^0.2.17",
     "@modelcontextprotocol/sdk": "^1.29.0",
     "trpc-cli": "^0.12.1",
-    "ts-morph": "^27.0.2",
-    "yaml": "^2.9.0",
     "zod": "4.4.3"
   },
   "devDependencies": {
