+++ apps/web/src/components/providers.tsx
@@ -1,29 +1,14 @@
-
-import { ConvexProvider, ConvexReactClient } from "convex/react";
-
 import { Toaster } from "@/components/ui/sonner";
 import { ThemeProvider } from "@/lib/theme";
 
-const convexUrl = import.meta.env.VITE_CONVEX_URL;
-const convex = convexUrl ? new ConvexReactClient(convexUrl) : null;
-
+// Convex is intentionally NOT wired up here: no routed page consumes it today,
+// and the client SDK is too heavy for the app entry chunk. If an analytics
+// page is (re)added, wrap that route in its own lazily-loaded ConvexProvider.
 export default function Providers({ children }: { children: React.ReactNode }) {
-  // If Convex is not configured, render without ConvexProvider
-  if (!convex) {
-    return (
-      <ThemeProvider>
-        {children}
-        <Toaster />
-      </ThemeProvider>
-    );
-  }
-
   return (
     <ThemeProvider>
-      <ConvexProvider client={convex}>
-        {children}
-        <Toaster />
-      </ConvexProvider>
+      {children}
+      <Toaster />
     </ThemeProvider>
   );
 }
