+++ apps/web/src/lib/docs/search-data.ts
@@ -1,10 +1,22 @@
 import { buildSearchSections, type SearchSection } from "./search";
-import { getAllPages } from "./source";
+import { getAllPages, loadAllRawPages } from "./source";
 
-export const searchSections: SearchSection[] = buildSearchSections(
-  getAllPages().map((page) => ({
-    url: page.url,
-    rawSource: page.raw,
-    frontmatter: page.frontmatter,
-  })),
-);
+let sectionsPromise: Promise<SearchSection[]> | null = null;
+
+/**
+ * Build the search index data on demand: the raw markdown of every docs page
+ * is loaded lazily (one fetch per page chunk, memoized) instead of being
+ * bundled into the client eagerly.
+ */
+export function loadSearchSections(): Promise<SearchSection[]> {
+  sectionsPromise ??= loadAllRawPages().then((rawByFilePath) =>
+    buildSearchSections(
+      getAllPages().map((page) => ({
+        url: page.url,
+        rawSource: rawByFilePath.get(page.filePath) ?? "",
+        frontmatter: page.frontmatter,
+      })),
+    ),
+  );
+  return sectionsPromise;
+}
