+++ apps/web/src/routes/docs/index.tsx
@@ -2,7 +2,7 @@ import { createFileRoute, notFound } from "@tanstack/react-router";
 
 import { DocsPageContent } from "@/components/docs/docs-page";
 import { docsPageHead } from "@/lib/docs/seo";
-import { getNeighbors, getPage } from "@/lib/docs/source";
+import { getNeighbors, getPage, preloadDocPageContent } from "@/lib/docs/source";
 
 /**
  * Exact match for `/docs` — renders the docs index page (`content/docs/index.mdx`).
@@ -15,6 +15,7 @@ export const Route = createFileRoute("/docs/")({
   loader: () => {
     const page = getPage([]);
     if (!page) throw notFound();
+    preloadDocPageContent(page.slug);
     return {
       slug: page.slug,
       frontmatter: page.frontmatter,
