+++ apps/web/src/routes/guides/$.tsx
@@ -2,13 +2,14 @@ import { createFileRoute, notFound } from "@tanstack/react-router";
 
 import { GuidePageContent } from "@/components/guides/guide-page";
 import { guidePageHead } from "@/lib/guides/seo";
-import { getGuidePage } from "@/lib/guides/source";
+import { getGuidePage, preloadGuidePageContent } from "@/lib/guides/source";
 
 export const Route = createFileRoute("/guides/$")({
   loader: ({ params }) => {
     const slug = (params._splat ?? "").split("/").filter(Boolean);
     const page = getGuidePage(slug);
     if (!page) throw notFound();
+    preloadGuidePageContent(page.slug);
     return {
       slug: page.slug,
       frontmatter: page.frontmatter,
