+++ apps/web/src/routes/guides/index.tsx
@@ -2,12 +2,13 @@ import { createFileRoute, notFound } from "@tanstack/react-router";
 
 import { GuidePageContent } from "@/components/guides/guide-page";
 import { guidePageHead } from "@/lib/guides/seo";
-import { getGuidePage } from "@/lib/guides/source";
+import { getGuidePage, preloadGuidePageContent } from "@/lib/guides/source";
 
 export const Route = createFileRoute("/guides/")({
   loader: () => {
     const page = getGuidePage([]);
     if (!page) throw notFound();
+    preloadGuidePageContent(page.slug);
     return {
       frontmatter: page.frontmatter,
     };
